Use UTCTime for commits' time and related changes
Updated Commit record, and other records that have an embedded commit in
them, so that the commit's time is in UTCTime instead of Text, and that
it is converted to ISO when sent to Elastic Search.

Changes include:

 - Updated all records to use UTCTime instead of the previously used
   Text.
 - Updated the functions that get a commit's time from git to convert it
   to UTCTime.
 - Created a GitCommit record that holds a commit as fetched e.g. a hash
   and a time. This was so that we don't create EmIssueCommit records
   instead in cases where there is no intention of embedding the commit
   into Issue.
 - Replaced the official Aeson release with a patched version that
   converts UTCTime to ISO with subseconds (".SSSZ") at the end, even in
   when there is no fractional part in the time.
 - Updated the Elastic Search mappings to send commits' times in the
   'basic_time' format.

Another major change, not originally intended to be part of this patch,
was the addition of an IssueProperties record that holds the properties
of an Issue. This allowed to make the builds without having to create
fake Issue records, like it was done before, e.g. Issue records that
only hold the properties and the rest of the fields are given 'empty'
values. This was no longer possible since we cannot create an empty
EmIssueCommit record for the Issue's 'commit' field because it's 'time'
field that is now a UTCTime field cannot be empty.

As a side-effect of that, the properties of an Issue are now stored as
an embedded IssueProperties record and stored as a 'nested' field in
Elastic Search. This was the preferred design that would be implemented
later, as it allows to separate properties (that may be customised) from
the rest of the core Issue fields.

Other side effects of the work described above included in the patch are:

 - Updated the versions of Aeson and Bloodhound packages used.
 - Updated the Stackage resolver version.
 - Moved all Elastic Search mappings into separate files, one per
   record.
 - Cleaned up various functions that are no longer in use.
 - Starting replacing usage of String with Text in the Parser and Issue
   records.
 - Made all Issue property fields optional (using Maybe), apart from the
   'title' property. They are also no longer stored in Elastic Search if
   the don't have a value.
 - Made coding style of Elastic Search mappings for records more
   readable.
 - Updated record mappings to list fields in alphabetical order.
 - Minor code style improvements and fixes.