Updated Build functions to work with IssueProperties
Build conditions and rules-related functions were previously working
with Issue records. They are now updated to work with IssueProperties
records, since at the stage in the program where the builds are
calculated we have not yet created Issue records - we only have the
issues as the annotations contained in the commits contents.