Added delay when updating heads
If a head commit was just analysed before trying to set it as the head
commit of a branch, project or repository, it may not be available yet
due to Elastic Search's near real time feature. We add here 1s of delay
to make sure that there is enough time for the commit to become
available to search.